rename mux package to http for more consistency
