+++ test/e2e/e2e_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/gavv/httpexpect"
 	"github.com/pressly/goose/v3"
 	"github.com/soggycactus/sharechat.dev/sharechat"
+	sharechathttp "github.com/soggycactus/sharechat.dev/sharechat/http"
 	"github.com/soggycactus/sharechat.dev/sharechat/memory"
-	"github.com/soggycactus/sharechat.dev/sharechat/mux"
 	"github.com/soggycactus/sharechat.dev/sharechat/postgres"
 	"github.com/soggycactus/sharechat.dev/sharechat/redis"
 )
@@ -129,7 +129,7 @@ func TestMemory(t *testing.T) {
 		Queue:       redis.NewQueue("0.0.0.0:6379", "", ""),
 	})
 
-	server := httptest.NewServer(mux.NewServer(controller).Handler)
+	server := httptest.NewServer(sharechathttp.NewServer(controller).Handler)
 	defer server.Close()
 	e := httpexpect.New(t, server.URL)
 	ServeNewRoom(e)
@@ -159,7 +159,7 @@ func TestServeNewRoom(t *testing.T) {
 		Queue:       redis.NewQueue("0.0.0.0:6379", "", ""),
 	})
 
-	server := httptest.NewServer(mux.NewServer(controller).Handler)
+	server := httptest.NewServer(sharechathttp.NewServer(controller).Handler)
 	defer server.Close()
 	e := httpexpect.New(t, server.URL)
 	ServeNewRoom(e)
@@ -189,7 +189,7 @@ func TestServeExistingRoom(t *testing.T) {
 		Queue:       redis.NewQueue("0.0.0.0:6379", "", ""),
 	})
 
-	server := httptest.NewServer(mux.NewServer(controller).Handler)
+	server := httptest.NewServer(sharechathttp.NewServer(controller).Handler)
 	defer server.Close()
 	e := httpexpect.New(t, server.URL)
